requirements.txt
celery==5.2.7
django-celery-results==2.5.1

Django==4.2.4
django-redis==5.3.0

djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==20.1.0
Pillow==9.4.0
django-cors-headers==3.13.0
redis==4.4.0
Markdown==3.4.1
django-filter==22.1
coreapi==2.3.3


